# api/openapi.yaml
openapi: 3.0.3
info:
  title: identity
  version: "1"
  description: |
    Who somebody is, once, for every service operated together. Sign in with
    GitHub or Discord, get one account, carry it everywhere.

    This file is the contract. `internal/api` implements it and a test holds
    the two equal: every route is described here, every path here is a route,
    and every response the test suite provokes is validated against these
    schemas. The Go client in this package is generated from it; a TypeScript
    client would be generated from the same file.

    Two kinds of credential exist and the difference is the security model.
    An **account token** comes from signing in here (or from `POST /v1/tokens`)
    and may manage the account. An **application token** comes from
    `POST /v1/exchange` and is bound to one application's origin (its
    `token.audience`); it may call `/v1/whoami` and `/v1/avatar`, revoke
    itself, sign its browser out with `/v1/signout`, and nothing else. A
    browser-facing service must hold only its own application tokens,
    server-side, and must check `token.audience` against its own origin
    before trusting a `whoami` answer.

    An application token minted for a browser that signed in here is linked
    to that sign-in: revoking the sign-in revokes the token, and signing in to
    the same application again from the same sign-in replaces it.
servers:
  - url: /
...
security:
  - bearer: []
paths:
...
  /signin/github/start:
    post:
      operationId: githubStart
      tags: [session]
      summary: Begin the GitHub device flow.
      description: |
        Called with a bearer token, the flow links the proved GitHub identity
        to that account instead of signing in.
      security:
        - {}
        - bearer: []
      responses:
        "200":
          description: A device code for the person to enter at GitHub.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/Throttled" }
        "501": { $ref: "#/components/responses/Unconfigured" }
        "502": { $ref: "#/components/responses/ProviderDown" }
  /signin/github/finish:
    post:
      operationId: githubFinish
      tags: [session]
      summary: Poll the GitHub device flow.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_code]
              properties:
                device_code: { type: string }
      responses:
        "201":
          description: "Signed in. With `X-Identity-Browser: 1` the token is set as a cookie and omitted from the body."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Token" }
        "200":
          description: The flow was a link, and it completed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Linked" }
        "202":
          description: Not finished at GitHub yet. Poll again after the interval, longer on slow_down.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [pending, slow_down] }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Failure" }
        "501": { $ref: "#/components/responses/Unconfigured" }
        "502": { $ref: "#/components/responses/ProviderDown" }
  /signin/discord/start:
    post:
      operationId: discordStart
      tags: [session]
      summary: Begin the Discord redirect flow.
      description: |
        Answers with the URL to send the browser to, and sets a state cookie
        the callback must match. Called with a bearer token, the flow links
        instead of signing in.
      security:
        - {}
        - bearer: []
      responses:
        "200":
          description: Where to send the browser.
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url: { type: string, format: uri }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/Throttled" }
        "500": { $ref: "#/components/responses/Failure" }
        "501": { $ref: "#/components/responses/Unconfigured" }
  /signin/discord/callback:
    get:
      operationId: discordCallback
      tags: [session]
      summary: Where Discord sends the browser back.
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
      responses:
        "303":
          description: Signed in or linked; the session cookie is set and the browser is sent to the page.
        "200":
          description: The sign-in failed, explained as a page.
          content:
            text/html:
              schema: { type: string }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer bsid_...`"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: "What went wrong, in a sentence." }
    AccountID:
      type: string
      description: A random 12-character hex id. Stable for the life of the account.
      pattern: "^[0-9a-f]{12}$"
      example: 3f9c2a81d04b
...
    Device:
      type: object
      required: [device_code, user_code, verification_uri, expires_in, interval]
      properties:
        device_code: { type: string, description: Opaque; send it to finish. }
        user_code: { type: string, description: What the person types at GitHub. }
        verification_uri: { type: string, format: uri }
        expires_in: { type: integer, description: Seconds. }
        interval: { type: integer, description: Minimum seconds between polls. }
    Token:
      type: object
      description: A freshly minted credential. The token itself is shown exactly once.
      required: [token, account, handle]
      properties:
        token: { type: string, description: "`bsid_<id>_<secret>`. Empty when set as a cookie instead." }
        account: { $ref: "#/components/schemas/AccountID" }
        handle: { type: string }
        expires_at: { type: string, format: date-time }
...
    Linked:
      type: object
      required: [linked, provider, handle]
      properties:
        linked: { type: boolean, enum: [true] }
        provider: { type: string, enum: [github, discord] }
        handle: { type: string }
...
  responses:
    BadRequest:
      description: The request was malformed; the error says how.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No live bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The token exists but may not do this; the error says why.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: It already exists, or belongs to somebody else.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Throttled:
      description: Too many attempts; Retry-After says when.
      headers:
        Retry-After:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Failure:
      description: This service could not do it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unconfigured:
      description: That provider or feature is not configured on this deployment.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ProviderDown:
      description: GitHub, Discord or the asset store could not be reached.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }